pyproject.toml
[project]
name = "cis-web-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask-cors>=5.0.0",
    "flask>=3.1.0",
    "google-generativeai>=0.8.4",
    "google-genai>=1.11.0",
]
